updated filters. added pub sub node
