+++ src/preprocess/CMakeLists.txt
@@ -1,3 +1,8 @@
-add_executable(${PROJECT_NAME}_node preprocess_node.cpp whitebalance.cpp median_subtract.cpp)
-add_dependencies(${PROJECT_NAME}_node ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-target_link_libraries(${PROJECT_NAME}_node ${catkin_LIBRARIES} ${OpenCV_LIBRARIES})
+add_executable(preprocess_node preprocess_node.cpp filters/bilateral_filter.cpp filters/whitebalance.cpp filters/median_subtract.cpp)
+add_dependencies(preprocess_node ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+target_link_libraries(preprocess_node ${catkin_LIBRARIES} ${OpenCV_LIBRARIES})
+
+
+add_executable(preprocess_test preprocess_test.cpp filters/bilateral_filter.cpp filters/whitebalance.cpp filters/median_subtract.cpp)
+add_dependencies(preprocess_test ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+target_link_libraries(preprocess_test ${catkin_LIBRARIES} ${OpenCV_LIBRARIES})
