+++ src/preprocess/filters/bilateral_filter.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <opencv/cv.h>
+
+cv::Mat bilateral_filter(cv::Mat &img);
