+++ src/preprocess/filters/median_subtract.cpp
