+++ src/preprocess/filters/median_subtract.h
