+++ src/preprocess/filters/whitebalance.cpp
@@ -1,4 +1,4 @@
-#include "preprocess/whitebalance.h"
+#include "preprocess/filters/whitebalance.h"
 #include <opencv2/core/core.hpp>
 #include <opencv2/imgproc.hpp>
 #include <vector>
