+++ src/preprocess/filters/whitebalance.h
